+++ launcher/src/bootstrap.c
@@ -12,6 +12,7 @@
 
 #include "bootstrap.h"
 #include "patchelf.h"
+#include "crash-report.h"
 
 // Get All Mods In Folder
 static void load(char **ld_preload, char *folder) {
@@ -97,12 +98,11 @@ static void exit_handler(__attribute__((unused)) int signal_id) {
 
 // Pre-Bootstrap
 void pre_bootstrap(int argc, char *argv[]) {
+    // Set Debug Tag
+    reborn_debug_tag = "(Launcher) ";
+
     // Disable stdout Buffering
     setvbuf(stdout, NULL, _IONBF, 0);
-
-    // Set Default Native Component Environment
-#define set_variable_default(name) set_and_print_env("MCPI_NATIVE_" name, getenv(name));
-    for_each_special_environmental_variable(set_variable_default);
 
     // Print Version
     for (int i = 1; i < argc; i++) {
@@ -114,29 +114,24 @@ void pre_bootstrap(int argc, char *argv[]) {
         }
     }
 
+    // Setup Logging
+    setup_log_file();
+
+    // --debug
+    for (int i = 1; i < argc; i++) {
+        if (strcmp(argv[i], "--debug") == 0) {
+            set_and_print_env("MCPI_DEBUG", "1");
+            break;
+        }
+    }
+
+    // Set Default Native Component Environment
+#define set_variable_default(name) set_and_print_env("MCPI_NATIVE_" name, getenv(name));
+    for_each_special_environmental_variable(set_variable_default);
+
     // GTK Dark Mode
 #ifndef MCPI_SERVER_MODE
     set_and_print_env("GTK_THEME", "Adwaita:dark");
-#endif
-
-    // Debug Zenity
-#ifndef MCPI_SERVER_MODE
-    {
-        const char *is_debug = getenv("MCPI_DEBUG");
-        if (is_debug != NULL && strlen(is_debug) > 0) {
-            set_and_print_env("ZENITY_DEBUG", "1");
-        }
-    }
-#endif
-
-    // AppImage
-#ifdef MCPI_IS_APPIMAGE_BUILD
-    {
-        char *owd = getenv("OWD");
-        if (owd != NULL && chdir(owd) != 0) {
-            ERR("AppImage: Unable To Fix Current Directory: %s", strerror(errno));
-        }
-    }
 #endif
 
     // Get Binary Directory
@@ -161,6 +156,19 @@ void pre_bootstrap(int argc, char *argv[]) {
 
     // Free Binary Directory
     free(binary_directory);
+
+    // Setup Crash Reports
+    setup_crash_report();
+
+    // AppImage
+#ifdef MCPI_IS_APPIMAGE_BUILD
+    {
+        char *owd = getenv("OWD");
+        if (owd != NULL && chdir(owd) != 0) {
+            ERR("AppImage: Unable To Fix Current Directory: %s", strerror(errno));
+        }
+    }
+#endif
 
     // Install Signal Handlers
     struct sigaction act_sigint;