+++ launcher/src/crash-report.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void setup_log_file();
+void setup_crash_report();
+
+#ifdef __cplusplus
+}
+#endif